+++ Clase Springboot/src/main/java/com/uade/tpo/demo/service/ProductoService.java
@@ -1,5 +1,5 @@
 package com.uade.tpo.demo.service;
-
+import java.math.BigDecimal;
 import com.uade.tpo.demo.entity.Producto;
 import com.uade.tpo.demo.exceptions.ImagenProductoNotFoundException;
 import com.uade.tpo.demo.exceptions.ProductoNotFoundException;
@@ -8,12 +8,12 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ProductoService {
-    List<Producto> getProductos(String nombre, Long categoriaId, Float precioMin, Float precioMax);
+    List<Producto> getProductos(String nombre, Long categoriaId, BigDecimal precioMin, BigDecimal precioMax);
     Optional<Producto> getProductoById(Long id);
 
-    Producto createProducto(Long categoriaId, String nombre, String descripcion, Float precio, Integer stock, Float descuento, List<String> imagenes);
+    Producto createProducto(Long categoriaId, String nombre, String descripcion, BigDecimal precio, Integer stock, BigDecimal descuento, List<String> imagenes);
 
-    Producto updateProducto(Long id, Long categoriaId, String nombre, String descripcion, Float precio, Integer stock, Float descuento) throws ProductoNotFoundException;
+    Producto updateProducto(Long id, Long categoriaId, String nombre, String descripcion, BigDecimal precio, Integer stock, BigDecimal descuento) throws ProductoNotFoundException;
 
     Producto actualizarStock(Long id, Integer nuevoStock) throws ProductoNotFoundException;
